Restructure repo, refactor styling (#7)
* Separated out helper functions into lib/. Noted todo items for cleanup

* Split up stylesheets

* Update colors

* Update colors and fix Airtable discrepancies

*Fix goBack issue
